+++ src/app/(withheader-footer)/contact-us/page.tsx
@@ -1,40 +1,52 @@
-import type { Metadata } from 'next';
+import type { Metadata } from "next";
 
-export const metadata: Metadata = {
-  title: 'About Us',
-  description:
-    'Hilop',
+// ✅ Dynamic metadata
+export async function generateMetadata(): Promise<Metadata> {
+  const pageTitle = "Contact Hilop – Get in Touch with Our Support Team";
+  const pageDescription =
+    "Have questions about Hilop? Contact our customer support for queries, feedback, or product details.";
 
-  openGraph: {
-    title: 'About Us',
-    description:
-      'hilop',
-    images: [
-      {
-        url: '',
-        width: 1200,
-        height: 630,
-        alt: 'About Hilop',
-      },
+  return {
+    title: pageTitle,
+    description: pageDescription,
+    keywords: [
+      "Hilop Contact",
+      "Customer Support Hilop",
+      "Hilop Helpline",
+      "Weight Loss Product Help",
     ],
-  },
+    alternates: {
+      canonical: "https://hilop.com/contact-us", 
+    },
+    openGraph: {
+      title: pageTitle,
+      description: pageDescription,
+      url: "https://hilop.com/contact",
+      siteName: "Hilop",
+      images: [
+        {
+          url: "/image/about-us/mission.webp", // replace with a "Contact" banner image if available
+          width: 1200,
+          height: 630,
+          alt: "Contact Hilop",
+        },
+      ],
+      type: "website",
+      locale: "en_US",
+    },
+    twitter: {
+      card: "summary_large_image",
+      title: pageTitle,
+      description: pageDescription,
+      images: ["/image/about-us/mission.webp"],
+      creator: "@hilop",
+      site: "@hilop",
+    },
+  };
+}
 
-  twitter: {
-    title: 'About Us',
-    description:
-      'Hilop',
-    images: [
-      {
-        url: '/image/about-us/mission.webp',
-        width: 1200,
-        height: 630,
-        alt: 'About Hilop',
-      },
-    ],
-  },
-};
+import BookCall from "./BookCallContent";
 
-import BookCall from './BookCallContent';
 export default function BookCallPage() {
   return <BookCall />;
 }
